[#3810] s3 refactor to prepare for new section
